chore(gridsome): Add comment about workbox webpack plugin
Gridsome v0.7.0 creates a styles js chunk during the webpack compilation. Such file is removed when css.split!==true but after the compilation finished.

As the workbox plugin considered that chunk as a precaching asset, the browser does not find it and throws an error. To prevent this issue, we exclude the file.